Created single function for both create and update calls
