+++ lib/fog/azurerm/network.rb
@@ -21,12 +21,11 @@ module Fog
       request :delete_virtual_network
       request :list_virtual_networks
       request :check_for_virtual_network
-      request :create_public_ip
+      request :create_or_update_public_ip
       request :delete_public_ip
       request :get_public_ip
       request :list_public_ips
       request :check_for_public_ip
-      request :update_public_ip
       request :create_subnet
       request :attach_network_security_group_to_subnet
       request :detach_network_security_group_from_subnet
